behold_agent/pyproject.toml
[project]
name = "behold-agent"
version = "0.1.0"
description = "WhatsApp Shopify agent with Google ADK integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]
